submissions/aadyant-sood/level-6/requirements.txt
altair==6.1.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.1.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.50
h11==0.16.0
httptools==0.7.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.21.0
neo4j==6.2.0
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
plotly==6.7.0
protobuf==7.34.1
pyarrow==24.0.0
pydeck==0.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.28
pytz==2026.2
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
starlette==1.0.0
streamlit==1.57.0
tenacity==9.1.4
toml==0.10.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
uvicorn==0.46.0
watchdog==6.0.0
websockets==16.0